Safari-609.1.11.1/LayoutTests/fast/events/AddEventListenerOptions-once-recursive-expected.txt

Tests support for 'once' member in AddEventListenerOptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


document.body.addEventListener('test', listenerFunction, { 'once': true })
PASS listenerCallCount is 0
document.body.dispatchEvent(new Event('test'))
PASS listenerCallCount is 1
document.body.dispatchEvent(new Event('test'))
PASS listenerCallCount is 1
PASS successfullyParsed is true

TEST COMPLETE
